+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/ImageService.java
@@ -6,6 +6,6 @@ public interface ImageService {
 
     public Optional<byte[]> findById(Long id);
 
-    public Integer insertImage(byte[] img);
+    public Long insertImage(byte[] img);
 
 }
