+++ services/src/main/java/ar/edu/itba/paw/services/TournamentServiceImpl.java
@@ -60,7 +60,7 @@ public class TournamentServiceImpl implements TournamentService {
 
     @Override
     public Tournament create(Long creator_id, String name, Long game_id, Region region, Elo elo, LocalDate start_date, LocalDate end_date, String format, Structure structure, Integer max_participants, byte[] image, Boolean openInscriptions, Boolean isFinished) {
-        Integer image_id = imageDao.insertImage(image);
+        Long image_id = imageDao.insertImage(image);
         return tournamentDao.create(creator_id, name, game_id, region, elo, start_date, end_date, format, structure, max_participants, image_id, openInscriptions, isFinished);
     }
